# plugins/gin/testdata/server/docs/openapi.json
{
    "components": {
        "schemas": {
            "github.com_gin-gonic_gin.Param": {
                "description": "is a single URL parameter, consisting of a key and a value.",
                "properties": {
                    "Key": {
                        "type": "string"
                    },
                    "Value": {
                        "type": "string"
                    }
                },
                "title": "GinParam",
                "type": "object"
            },
            "github.com_gin-gonic_gin.Params": {
                "description": "is a Param-slice, as returned by the router.\n\nThe slice is ordered, the first URL parameter is also the first slice value.\n\nIt is therefore safe to read values by the index.",
                "items": {
                    "$ref": "#/components/schemas/github.com_gin-gonic_gin.Param"
                },
                "title": "GinParams",
                "type": "array"
            },
            "server_pkg_view.ErrCode": {
                "description": "\u003ctable\u003e\u003ctr\u003e\u003cth\u003eValue\u003c/th\u003e\u003cth\u003eKey\u003c/th\u003e\u003cth\u003eDescription\u003c/th\u003e\u003c/tr\u003e\u003ctr\u003e\u003ctd\u003e10000\u003c/td\u003e\u003ctd\u003eCodeNotFound\u003c/td\u003e\u003ctd\u003eResource not found\u003c/td\u003e\u003c/tr\u003e\u003ctr\u003e\u003ctd\u003e10001\u003c/td\u003e\u003ctd\u003eCodeCancled\u003c/td\u003e\u003ctd\u003eRequest canceld\u003c/td\u003e\u003c/tr\u003e\u003ctr\u003e\u003ctd\u003e10002\u003c/td\u003e\u003ctd\u003eCodeUnknown\u003c/td\u003e\u003ctd\u003e\u003c/td\u003e\u003c/tr\u003e\u003ctr\u003e\u003ctd\u003e10003\u003c/td\u003e\u003ctd\u003eCodeInvalidArgument\u003c/td\u003e\u003ctd\u003e\u003c/td\u003e\u003c/tr\u003e\u003c/table\u003e",
                "enum": [
                    10000,
                    10001,
                    10002,
                    10003
                ],
                "title": "ViewErrCode",
                "type": "integer"
            },
            "server_pkg_view.Error": {
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/server_pkg_view.ErrCode"
                    },
                    "msg": {
                        "type": "string"
                    }
                },
                "title": "ViewError",
                "type": "object"
            },
            "server_pkg_view.GoodsCreateReq": {
                "properties": {
                    "cover": {
                        "description": "封面图",
                        "type": "string"
                    },
                    "images": {
                        "description": "详情图",
                        "items": {
                            "$ref": "#/components/schemas/server_pkg_view.Image"
                        },
                        "type": "array"
                    },
                    "price": {
                        "description": "价格(分)",
                        "type": "integer"
                    },
                    "subTitle": {
                        "description": "商品描述",
                        "type": "string"
                    },
                    "title": {
                        "description": "商品标题",
                        "type": "string"
                    }
                },
                "required": [
                    "title",
                    "price"
                ],
                "title": "ViewGoodsCreateReq",
                "type": "object"
            },
            "server_pkg_view.GoodsCreateRes": {
                "properties": {
                    "Status": {
                        "description": "测试引用第三方包",
                        "$ref": "#/components/schemas/github.com_gin-gonic_gin.Params"
                    },
                    "guid": {
                        "description": "商品 GUID",
                        "type": "string"
                    },
                    "raw": {
                        "deprecated": true,
                        "description": "测试引用内置包类型",
                        "type": "object"
                    },
                    "selfRef": {
                        "description": "测试循环引用",
                        "$ref": "#/components/schemas/server_pkg_view.SelfRefType"
                    },
                    "stringAlias": {
                        "description": "测试类型别名",
                        "type": "string"
                    }
                },
                "title": "ViewGoodsCreateRes",
                "type": "object"
            },
            "server_pkg_view.GoodsDownRes": {
                "properties": {
                    "Status": {
                        "type": "string"
                    }
                },
                "title": "ViewGoodsDownRes",
                "type": "object"
            },
            "server_pkg_view.GoodsInfoRes": {
                "properties": {
                    "cover": {
                        "type": "string"
                    },
                    "mapInt": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/server_pkg_view.Property"
                        },
                        "type": "object"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "properties": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/server_pkg_view.Property"
                        },
                        "type": "object"
                    },
                    "subTitle": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "title": "ViewGoodsInfoRes",
                "type": "object"
            },
            "server_pkg_view.Image": {
                "description": "商品图片",
                "properties": {
                    "url": {
                        "description": "图片链接",
                        "type": "string"
                    }
                },
                "required": [
                    "url"
                ],
                "title": "ViewImage",
                "type": "object"
            },
            "server_pkg_view.Property": {
                "properties": {
                    "title": {
                        "type": "string"
                    }
                },
                "title": "ViewProperty",
                "type": "object"
            },
            "server_pkg_view.SelfRefType": {
                "properties": {
                    "data": {
                        "type": "string"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/server_pkg_view.SelfRefType"
                    }
                },
                "title": "ViewSelfRefType",
                "type": "object"
            }
        },
        "securitySchemes": {
            "oauth2": {
                "flows": {
                    "implicit": {
                        "authorizationUrl": "https://example.org/api/oauth/dialog",
                        "scopes": {
                            "goods:write": "modify pets in your account",
                            "read:pets": "read your pets"
                        }
                    }
                },
                "type": "oauth2"
            }
        }
    },
    "info": {
        "description": "Example description for Example",
        "title": "This is an Example",
        "version": ""
    },
    "openapi": "3.1.0",
    "paths": {
        "/api/goods": {
            "post": {
                "description": "创建商品接口",
                "operationId": "shop.GoodsCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/server_pkg_view.GoodsCreateReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server_pkg_view.GoodsCreateRes"
                                }
                            }
                        },
                        "description": "创建成功"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server_pkg_view.Error"
                                }
                            }
                        },
                        "description": "参数无效"
                    }
                },
                "summary": "创建商品",
                "tags": [
                    "Goods"
                ]
            }
        },
        "/api/goods/{guid}": {
            "delete": {
                "description": "删除商品",
                "operationId": "shop.GoodsDelete",
                "parameters": [
                    {
                        "in": "query",
                        "name": "formDataField",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "oauth2": [
                            "goods:write"
                        ]
                    }
                ],
                "summary": "删除商品",
                "tags": [
                    "High Priority Tag"
                ]
            }
        },
        "/api/goods/{guid}/down": {
            "post": {
                "description": "下架商品",
                "operationId": "shop.GoodsDown",
                "parameters": [
                    {
                        "description": "商品 GUID",
                        "in": "path",
                        "name": "guid",
                        "required": true,
                        "schema": {
                            "title": "guid",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Default Query",
                        "in": "query",
                        "name": "defaultQuery",
                        "schema": {
                            "default": "xxxx",
                            "title": "defaultQuery",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "dateRange": {
                                        "description": "日期范围",
                                        "items": {
                                            "type": "string"
                                        },
                                        "title": "dateRange",
                                        "type": "array"
                                    },
                                    "defaultPostForm": {
                                        "description": "Default Post Form",
                                        "title": "defaultPostForm",
                                        "type": "string"
                                    },
                                    "operatorUid": {
                                        "description": "操作人 UID",
                                        "title": "operatorUid",
                                        "type": "string"
                                    }
                                },
                                "title": "ShopGoodsDownRequest",
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/server_pkg_view.GoodsDownRes"
                                }
                            }
                        }
                    }
                },
                "summary": "下架商品",
                "tags": [
                    "Goods"
                ]
            }
        },
        "/api/v2/goods/{guid}": {
            "get": {
                "description": "商品详情",
                "operationId": "shop.GoodsInfo",
                "parameters": [
                    {
                        "in": "path",
                        "name": "guid",
                        "required": true,
                        "schema": {
                            "title": "guid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server_pkg_view.GoodsInfoRes"
                                }
                            }
                        }
                    }
                },
                "summary": "商品详情",
                "tags": [
                    "Goods"
                ]
            }
        },
        "/wrapped-handler": {
            "get": {
                "deprecated": true,
                "operationId": "shop.WrappedHandler",
                "parameters": [
                    {
                        "in": "query",
                        "name": "hello",
                        "schema": {
                            "title": "hello",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "world",
                        "schema": {
                            "title": "world",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "data": {
                                            "additionalProperties": {
                                                "description": "Any Type",
                                                "type": "object"
                                            },
                                            "type": "object"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "data",
                                        "msg"
                                    ],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "自定义响应函数"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Any Json Type",
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Shop"
                ]
            }
        }
    }
}